+++ NEWS.md
@@ -1,5 +1,7 @@
 # brandseyer2 0.0.1.9002
 
+* `account_list()` can now include inactive accounts.
+
 # brandseyer2 0.0.1.9001
 
 * Added a `NEWS.md` file to track changes to the package.